LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_attributes_touch-nonstandard-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_unreached: missing pointerover event Reached unreachable code
CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_unreached: missing pointerenter event Reached unreachable code
CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_unreached: missing pointerdown event Reached unreachable code
CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_unreached: missing pointerup event Reached unreachable code
CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_unreached: missing pointerout event Reached unreachable code
CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_unreached: missing pointerleave event Reached unreachable code
CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_unreached: missing pointermove event Reached unreachable code
CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_unreached: missing pointerover event Reached unreachable code
CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_unreached: missing pointerenter event Reached unreachable code
CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_unreached: missing pointerdown event Reached unreachable code
CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_unreached: missing pointerup event Reached unreachable code
CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_unreached: missing pointerout event Reached unreachable code
CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_unreached: missing pointerleave event Reached unreachable code
CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_unreached: missing pointermove event Reached unreachable code

done

Harness Error (FAIL), message = Unhandled rejection: assert_unreached: missing pointermove event Reached unreachable code

FAIL Test pointer events in the main document assert_unreached: missing pointerover event Reached unreachable code
FAIL Test pointer events in an iframe assert_unreached: missing pointerover event Reached unreachable code
